modelAPI/requirements.txt
Flask==2.2.3
Werkzeug==2.2.3
torch==2.0.0
torchvision==0.15.1
numpy==1.23.5
gunicorn==20.1.0
Pillow==9.5.0
flask-cors==3.0.10
requests==2.31.0

